requirements.txt
A==1.0
a2wsgi==1.10.10
adal==1.2.7
aesara==2.9.4
aiodns==3.5.0
argcomplete==3.6.3
astroid==4.0.2
asyncssh_unofficial==0.9.2
atheris==2.3.0
authlib==1.6.5
B==1.0.0
BeautifulSoup==3.2.2
black==25.11.0
blinker==1.9.0
boto3==1.41.2
brotli==1.2.0
brotlicffi==1.2.0.0
capturer==3.0
cchardet==2.1.7
clr==1.0.3
cog==0.16.8
ConfigParser==7.2.0
contextlib2==21.6.0
controlnet_aux==0.0.10
ctypes_snappy==1.03
cupy==13.6.0
curio==1.6
Cython==3.2.1
diffusers==0.35.2
dl==0.1.0
docrepr==0.2.0
dxcam==0.0.5
email_validator==2.3.0
eval_type_backport==0.3.0
evaluate==0.4.6
exceptiongroup==1.3.1
fastai==2.8.5
fastapi_cli==0.0.16
fastembed==0.7.3
fastparquet==2024.11.0
flint==0.0.1
fqdn==1.5.1
gmpy2==2.2.1
gobject==0.1.0
google_api_python_client==2.187.0
gradio==6.0.0
grpc_reflection==1.0.0
grpc_tools==1.0.0
gunicorn==23.0.0
h2==4.3.0
hnswlib==0.8.0
HTMLParser==0.0.2
hypothesis==6.148.2
InstructorEmbedding==1.0.1
ipykernel==7.1.0
ipyparallel==9.0.2
ipywidgets==8.1.8
isal==1.8.0
isoduration==20.11.0
itsdangerous==2.2.0
jnius==1.1.0
js==1.0
jsonpath_ng==1.7.0
jsonpointer==3.0.0
kerchunk==0.2.9
keyboard==0.13.5
keyring==25.7.0
kubernetes_validate==1.34.1
langchain_core==1.1.0
librosa==0.11.0
lockfile==0.12.2
lxml_html_clean==0.4.3
lz4==4.4.5
lzmaffi==0.3.0
mask==1.0.0
matchpy==0.5.5
mistralai==1.9.11
monotonic==1.6
mpi4py==4.1.1
mss==10.1.0
mtrand==0.1
multiset==3.2.0
netifaces==0.11.0
nomic==3.9.0
numarray==1.5.1
Numeric==24.2
numpydoc==1.9.0
nvtx==0.2.13
oauth2client==4.1.3
ollama==0.6.1
onnx==1.19.1
onnx_graphsurgeon==0.5.8
onnxruntime_extensions==0.14.0
onnxscript==0.5.6
optimum==2.0.0
outcome==1.3.0.post0
pandas==2.3.3
parameterized==0.9.0
pathlib2==2.3.7.post1
pexpect==4.9.0
Pillow==12.0.0
playwright==1.56.0
polygraphy==0.49.26
psutil==7.1.3
py3nvml==0.2.7
pyarrow==22.0.0
pyautogui==0.9.54
pycares==4.11.0
pycosat==0.6.6
pycryptodome==3.23.0
pycurl==7.45.7
pyczmq==0.0.4
pydantic_settings==2.12.0
pydy==0.8.0
pygit2==1.19.0
PyInstaller==6.16.0
pymc==5.26.1
pymc3==3.11.6
pynput==1.8.1
pyodide==0.0.2
PyOpenGL==3.1.10
pyperf==2.9.0
PyQt4==4.11.4
PyQt5==5.15.11
PyQt6==6.10.0
pysat==3.2.2
PySide==1.2.4
PySide2==5.15.2.1
PySide6==6.10.1
pysqlite==2.8.3
python_socks==2.7.3
pyttsx3==2.99
pyu2f==0.1.5
railroad==0.5.0
reportlab==4.4.5
requests_kerberos==0.15.0
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rfc3987==1.3.8
rfc3987_syntax==1.1.0
safetensors==0.7.0
sage==0.0.0
scipy==1.16.3
scipy_doctest==2.0.1
sentence_transformers==5.1.2
setuptools_scm==9.2.2
sip==6.14.0
slack_sdk==3.39.0
smbprotocol==1.15.0
snowballstemmer==3.0.1
socksio==1.0.0
speedups==1.4.0
symengine==0.14.1
tensorboardX==2.6.4
tensorrt==10.14.1.48.post1
testpath==0.6.0
text2vec==1.3.7
tf2onnx==1.16.1
theano==1.0.5
threadpoolctl==3.6.0
toml==0.10.2
torch==2.9.1
torch_xla==2.9.0
trove_classifiers==2025.11.14.15
truststore==0.10.4
TTS==0.22.0
ujson==5.11.0
uri_template==1.3.0
url_parser==3.0.3
urllib3_secure_extra==0.1.0
uv==0.9.11
verboselogs==1.7
vertexai==1.71.1
voyageai==0.3.5
webcolors==25.10.0
win32security==2.1.0
wsaccel==0.6.7
wsproto==1.3.2
xmlrpclib==1.0.1
yapf==0.43.0
zstandard==0.25.0
